refactor: remove unused re-exports and adjust imports
Eliminated unused re-exports from the task-details index file to streamline module structure. Updated all related imports to reference specific component paths, improving clarity and reducing ambiguity.
