--- apps/nextjs/src/components/features/task-details/index.ts
@@ -1,3 +1,0 @@
-export * from './task-detail-drawer';
-export * from './task-detail-modal';
-export * from './task-detail-side';
