+++ apps/nextjs/src/components/features/tasks/tasks-board/tasks-board-list-item/tasks-name/tasks-name-field.tsx
@@ -1,6 +1,6 @@
 import type React from 'react';
 import { memo, useCallback, useEffect, useMemo, useState } from 'react';
-import { useTaskDetailDrawerRef } from '@/components/features/task-details';
+import { useTaskDetailDrawerRef } from '@/components/features/task-details/task-detail-drawer';
 import { Flex } from '@/components/ui/flex';
 import type { InputProps } from '@/components/ui/input';
 import { InputText } from '@/components/ui/input-text';
