+++ apps/nextjs/src/components/features/tasks/tasks-files/use-tasks-files-detail.tsx
@@ -1,7 +1,7 @@
 import { useParams, usePathname } from 'next/navigation';
 import { useEffect } from 'react';
 import { useTaskDetail } from '@/components/features/task-detail';
-import { useTaskDetailModal } from '@/components/features/task-details';
+import { useTaskDetailModal } from '@/components/features/task-details/task-detail-modal';
 import type { Params } from '@/shared/nextjs/navigation';
 
 type Props = {
