+++ apps/nextjs/src/components/pages/inbox/components/archive/archive.tsx
@@ -9,7 +9,7 @@ import {
   InboxSkeleton,
   useInboxTaskDetail,
 } from '@/components/features/inbox';
-import { TaskDetailSide } from '@/components/features/task-details';
+import { TaskDetailSide } from '@/components/features/task-details/task-detail-side';
 import { TasksProvider } from '@/components/features/tasks';
 import { Flex } from '@/components/ui/flex';
 import { useInboxArchivePageQuery } from '@/hooks/queries/app';
