+++ apps/nextjs/src/components/pages/projects/components/calendar/calendar.tsx
@@ -1,5 +1,5 @@
 import { memo, useCallback } from 'react';
-import { TaskDetailModal } from '@/components/features/task-details';
+import { TaskDetailModal } from '@/components/features/task-details/task-detail-modal';
 import {
   TasksCalendar,
   TasksCalendarContent,
