+++ apps/nextjs/src/components/pages/projects/components/list/list.tsx
@@ -1,5 +1,5 @@
 import { memo, useCallback } from 'react';
-import { TaskDetailDrawer } from '@/components/features/task-details';
+import { TaskDetailDrawer } from '@/components/features/task-details/task-detail-drawer';
 import {
   AddTaskButton,
   CustomizeButton,
